requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.2
